separate 2 group for diff limiter rule
