+++ src/router/router.go
@@ -2,9 +2,7 @@ package router
 
 import (
 	"api"
-	"api/middlewares"
-	fixedwindow "api/middlewares/ratelimiter/fixedwindowcounter"
-	tokenbucket "api/middlewares/ratelimiter/tokenbucket"
+
 	"github.com/labstack/echo"
 )
 
@@ -12,22 +10,15 @@ func New() *echo.Echo {
 	e := echo.New()
 
 	// create groups
-	jwtGroup := e.Group("/jwt")
-
-	// set all middlewares
-	middlewares.SetMainMiddlewares(e)
-	middlewares.SetJwtMiddlewares(jwtGroup)
-
-	config := fixedwindow.NewConfig("userlimiter", 3, "minute")
-	configTB := tokenbucket.NewConfig("userlimiter", 3, "minute")
-	e.Use(fixedwindow.Limiter(config))
-	e.Use(tokenbucket.Limiter(configTB))
+	userGroup := e.Group("/v1.0/")
+	authGroup := e.Group("/v1.0/auth")
 
 	// set main routes
 	api.MainGroup(e)
 
 	// set group routes
-	api.JwtGroup(jwtGroup)
+	api.UserGroup(userGroup)
+	api.AuthGroup(authGroup)
 
 	return e
 }
